requirements.txt
flask==3.0.3
gunicorn==21.2.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.117.0
google-cloud-pubsub==2.27.1
langchain-anthropic==0.3.22
python-dotenv==1.0.1
twilio==9.3.8